OSIS-1104 fuse application fill tree version
