+++ program_management/ddd/command.py
@@ -741,28 +741,6 @@ class FillProgramTreeVersionContentFromProgramTreeVersionCommand(interface.Comma
 
 
 @attr.s(frozen=True, slots=True)
-class FillProgramTreeTransitionContentFromProgramTreeVersionCommand(interface.CommandRequest):
-    from_year = attr.ib(type=int)
-    from_offer_acronym = attr.ib(type=str)
-    from_version_name = attr.ib(type=str)
-    from_transition_name = attr.ib(type=str)
-    to_year = attr.ib(type=int)
-    to_offer_acronym = attr.ib(type=str)
-    to_version_name = attr.ib(type=str)
-    to_transition_name = attr.ib(type=str)
-    transaction_id = attr.ib(type=uuid.UUID, default=attr.Factory(uuid.uuid4))
-
-
-@attr.s(frozen=True, slots=True)
-class FillTreeVersionContentFromPastYearCommand(interface.CommandRequest):
-    to_year = attr.ib(type=int)
-    to_offer_acronym = attr.ib(type=str)
-    to_version_name = attr.ib(type=str)
-    to_transition_name = attr.ib(type=str)
-    transaction_id = attr.ib(type=uuid.UUID, default=attr.Factory(uuid.uuid4))
-
-
-@attr.s(frozen=True, slots=True)
 class CopyTreeCmsFromPastYear(interface.CommandRequest):
     code = attr.ib(type=str)
     year = attr.ib(type=int)
